gateway/requirements.txt
fastapi
uvicorn
requests
python-jose
python-dotenv
httpx